Added raceTimeout
